refactor: 도메인 객체 core 패키지로 이동
