+++ src/main/java/com/gyeryongbrother/pickandtest/dataaccess/adapter/StockRepositoryImpl.java
@@ -3,7 +3,7 @@ package com.gyeryongbrother.pickandtest.dataaccess.adapter;
 import com.gyeryongbrother.pickandtest.dataaccess.entity.StockEntity;
 import com.gyeryongbrother.pickandtest.dataaccess.mapper.StockDataAccessMapper;
 import com.gyeryongbrother.pickandtest.dataaccess.repository.StockJpaRepository;
-import com.gyeryongbrother.pickandtest.domain.Stock;
+import com.gyeryongbrother.pickandtest.domain.core.Stock;
 import com.gyeryongbrother.pickandtest.domain.service.ports.output.StockRepository;
 import lombok.RequiredArgsConstructor;
