+++ src/main/java/com/gyeryongbrother/pickandtest/domain/core/MarketCapitalization.java
@@ -1,4 +1,4 @@
-package com.gyeryongbrother.pickandtest.domain;
+package com.gyeryongbrother.pickandtest.domain.core;
 
 import java.time.LocalDate;
