+++ src/main/java/com/gyeryongbrother/pickandtest/domain/service/ports/output/StockRepository.java
@@ -1,6 +1,6 @@
 package com.gyeryongbrother.pickandtest.domain.service.ports.output;
 
-import com.gyeryongbrother.pickandtest.domain.Stock;
+import com.gyeryongbrother.pickandtest.domain.core.Stock;
 
 public interface StockRepository {
