+++ src/test/java/com/gyeryongbrother/pickandtest/dataaccess/adapter/StockQueryRepositoryImplTest.java
@@ -9,9 +9,9 @@ import com.gyeryongbrother.pickandtest.dataaccess.config.TestQuerydslConfig;
 import com.gyeryongbrother.pickandtest.dataaccess.entity.DividendEntity;
 import com.gyeryongbrother.pickandtest.dataaccess.entity.StockEntity;
 import com.gyeryongbrother.pickandtest.dataaccess.entity.StockPriceEntity;
-import com.gyeryongbrother.pickandtest.domain.Dividend;
-import com.gyeryongbrother.pickandtest.domain.Stock;
-import com.gyeryongbrother.pickandtest.domain.StockPrice;
+import com.gyeryongbrother.pickandtest.domain.core.Dividend;
+import com.gyeryongbrother.pickandtest.domain.core.Stock;
+import com.gyeryongbrother.pickandtest.domain.core.StockPrice;
 import com.gyeryongbrother.pickandtest.domain.service.ports.output.StockQueryRepository;
 import jakarta.persistence.EntityManager;
 import java.math.BigDecimal;
